+++ PythonRobotics/PathPlanning/pathset_planning/deform_path_set.py
@@ -15,7 +15,7 @@ if __name__ == "__main__":
     sys.path.append(ROOT_DIR)
     os.chdir(ROOT_DIR)
 
-from visualizer import Visualizer
+from PythonRobotics.PathPlanning.pathset_planning.visualizer import Visualizer
 from world_map import CityMap, Block
 from cost_map import CityCostMapLayer
 from rrt import SimpleNode
